pyproject.toml
[project]
name = "state-space-playground"
version = "0.1.0"
description = "Sandbox for experimenting with state-space models on spatial bandit neural data."
readme = "README.md"
authors = [
    { name = "Eric Denovellis", email = "edenovellis@gmail.com" },
]
requires-python = ">=3.11,<3.13"
dependencies = [
    # Core scientific stack
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "xarray>=2023.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "networkx>=3.0",
    "tqdm>=4.65.0",

    # JAX with bundled CUDA 12 wheels (recommended per JAX docs)
    "jax[cuda12]>=0.4.4",

    # State-space model libraries
    "dynamax",
    "state_space_practice @ git+https://github.com/edeno/state_space_practice",

    # Non-local detector (shared with parent project)
    "non_local_detector @ git+https://github.com/LorenFrankLab/non_local_detector.git",

    # Data loader transitive dependencies (required by vendored data_loaders/)
    "spyglass-neuro @ git+https://github.com/LorenFrankLab/spyglass.git",
    "track_linearization",
    "ripple_detection",
    "datajoint",

    # NetCDF support for xarray result files
    "netcdf4>=1.6.0",
    "h5netcdf>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.0.0",
    "jupytext>=1.16.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# Vendored from continuum-swr-replay; keep as-is to minimize drift.
"src/state_space_playground/data_loaders/bandit_task.py" = ["E402", "F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"
